Fix: Add redirection endpoint to access after the user accepts the "Terms of Service"
Using the information from the Mantis report https://mantis.ilias.de/view.php?id=39629, we identified the correct way to fix the redirection bug, which occurs when the provider platform requires acceptance of usage terms.

Initially, we attempted to apply the fix only within the LTI-related files. However, we discovered that the logic also needs to be added in ilStartUpGUI, because the method $frontend->authenticate() sets the session variable orig_request_target as empty. Therefore, this is the only suitable place where the bug can be effectively resolved.