[LTI] Fix: user login fixed
Related with the test: https://testrail.ilias.de//index.php?/tests/view/79639

When the user selects the privacy option "External User ID combined with a unique ILIAS platform ID formatted as an email address," no user ID was sent. This change allows that ID to be sent.

Additionally, a bug related to permissions while an LTI consumer object is offline was detected and fixed.