Update maintenance.md – responsibility for UI-Service (#10537)
* Update maintenance.md – responsibility for UI-Service

Dear @ILIAS-eLearning/technical-board,  @matthiaskunkel, @thibsy,
as announced, I would like to propose taking responsibility for the “UI Service” component and the corresponding UI framework.

I hope to be able to make a contribution in this way, although many more contributions from the community will be required, especially on a continuous level.

Best regards,
@oliversamoila

* Update maintenance.md