fix: ilLTIAppEventListener changed
Due to the rename of the LTI module to LTIProvider the class dedicated to detect the events was obsolete. After the last updates of ILIAS 10, ILIAS is able to detect when one module is listening, but it doesn't have the proper listener class. For that reason, we include this new class in order to have again the listener ready. Without this change the LTI connection is impossible, due to the error that ILIAS raise.

This was detected in the report [0046456](https://mantis.ilias.de/bug_update_page.php)