fix: connections and learning progress
With these changes we fix several bugs related with the connection and the learning progress transference.
In the first hand, ILIAS wasn't able to save the correct LTI version in the DB. This was the reason of many wrong connections. Besides that LTI use a parameter called 'client_id' which has the same name as the `client_id` of ILIAS, for that reason we adapt the code to remove the need of include that parameter. In addition, we include the JWT field in the LTI configuration as mandatory. Because after this changes we were able to improve the security in the protocol, before it wasn't needed. In addition, we fix one lang variable that was missing.

In the other hand,we improve the method to determine the learning progress (status and percentage). Besides that, there was a bug in the status, when the user has a punctuation of 0, ILIAS wasn't able to know that the test was finished. In order to fix it we added a new field in the lti_consumer_results table.